+++ src/aks-preview/azext_aks_preview/custom.py
@@ -127,9 +127,7 @@ from azext_aks_preview.maintenanceconfiguration import (
 )
 
 from .vendored_sdks.azure_mgmt_preview_aks.v2022_05_02_preview.models import (
-    AgentPool,
     AgentPoolUpgradeSettings,
-    ContainerServiceStorageProfileTypes,
     CreationData,
     KubeletConfig,
     LinuxOSConfig,
@@ -693,6 +691,7 @@ def aks_create(
     nat_gateway_idle_timeout=None,
     outbound_type=None,
     network_plugin=None,
+    network_plugin_mode=None,
     network_policy=None,
     auto_upgrade_channel=None,
     cluster_autoscaler_profile=None,
@@ -789,6 +788,7 @@ def aks_create(
     disk_driver_version=None,
     disable_disk_driver=False,
     disable_file_driver=False,
+    enable_blob_driver=None,
     disable_snapshot_controller=False,
     enable_apiserver_vnet_integration=False,
     apiserver_subnet_id=None,
@@ -901,6 +901,8 @@ def aks_update(
     disable_disk_driver=False,
     enable_file_driver=False,
     disable_file_driver=False,
+    enable_blob_driver=None,
+    disable_blob_driver=None,
     enable_snapshot_controller=False,
     disable_snapshot_controller=False,
     enable_apiserver_vnet_integration=False,
@@ -2388,10 +2390,9 @@ def aks_pod_identity_add(cmd, client, resource_group_name, cluster_name,
         pod_identity.binding_selector = binding_selector
     pod_identities.append(pod_identity)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2418,10 +2419,9 @@ def aks_pod_identity_delete(cmd, client, resource_group_name, cluster_name,
                 continue
             pod_identities.append(pod_identity)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2451,10 +2451,9 @@ def aks_pod_identity_exception_add(cmd, client, resource_group_name, cluster_nam
         name=exc_name, namespace=exc_namespace, pod_labels=pod_labels)
     pod_identity_exceptions.append(exc)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2480,10 +2479,9 @@ def aks_pod_identity_exception_delete(cmd, client, resource_group_name, cluster_
                 continue
             pod_identity_exceptions.append(exc)
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,
@@ -2517,10 +2515,9 @@ def aks_pod_identity_exception_update(cmd, client, resource_group_name, cluster_
         raise CLIError(
             'pod identity exception {}/{} not found'.format(exc_namespace, exc_name))
 
-    from azext_aks_preview.decorator import AKSPreviewModels
-
+    from azext_aks_preview.managed_cluster_decorator import AKSPreviewManagedClusterModels
     # store all the models used by pod identity
-    pod_identity_models = AKSPreviewModels(
+    pod_identity_models = AKSPreviewManagedClusterModels(
         cmd, CUSTOM_MGMT_AKS_PREVIEW).pod_identity_models
     _update_addon_pod_identity(
         instance, enable=True,