+++ src/spring/azext_spring/commands.py
@@ -75,7 +75,7 @@ def load_command_table(self, _):
 
     gateway_route_config_cmd_group = CliCommandType(
         operations_tmpl='azext_spring.gateway#{}',
-        client_factory=cf_spring_20220101preview
+        client_factory=cf_spring_20220501preview
     )
 
     api_portal_cmd_group = CliCommandType(
@@ -248,8 +248,7 @@ def load_command_table(self, _):
 
     with self.command_group('spring gateway',
                             custom_command_type=gateway_cmd_group,
-                            exception_handler=handle_asc_exception,
-                            is_preview=True) as g:
+                            exception_handler=handle_asc_exception) as g:
         g.custom_show_command('show', 'gateway_show', table_transformer=transform_spring_cloud_gateway_output)
         g.custom_command('update', 'gateway_update', validator=validate_gateway_update, supports_no_wait=True)
         g.custom_command('clear', 'gateway_clear', supports_no_wait=True)
@@ -276,8 +275,7 @@ def load_command_table(self, _):
 
     with self.command_group('spring api-portal',
                             custom_command_type=api_portal_cmd_group,
-                            exception_handler=handle_asc_exception,
-                            is_preview=True) as g:
+                            exception_handler=handle_asc_exception) as g:
         g.custom_show_command('show', 'api_portal_show', table_transformer=transform_api_portal_output)
         g.custom_command('update', 'api_portal_update', validator=validate_api_portal_update)
         g.custom_command('clear', 'api_portal_clear')