+++ src/aks-preview/HISTORY.rst
@@ -12,12 +12,31 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+0.5.86
++++++++
+
+* Support network plugin mode for enabling Azure CNI Overlay preview feature.
+
+0.5.85
+++++++
+
+* Add support for Blob csi driver.
+
+0.5.84 (NOT RELEASED)
++++++++++++++++++++++
+
+* Skip this version due to conflict.
+
+0.5.83
+++++++
+
 * Update the minimum required cli core version to `2.37.0`.
 * Enable v2 decorator pattern.
 * Fix container name inconsistency for private clusters in kollect command.
+* Temp fix for properties missing in KMS profile in update scenario.
 
 0.5.82
-+++++++
+++++++
 
 * Support Key Vault with private link when enabling Azure KeyVault KMS.
 
@@ -75,8 +94,8 @@ Pending
 
 * Fix import issues with command group `az aks draft`
 
-0.5.72
-++++++
+0.5.72 (NOT RELEASED)
++++++++++++++++++++++
 
 * First public release for `az aks draft`