+++ src/spring/HISTORY.md
@@ -1,5 +1,14 @@
 Release History
 ===============
+1.1.2
+---
+* Support configure Germination Grace Period Seconds for deployments.
+* Fix the arguments parsing of the Command `az spring app create` with "--container-image".
+
+1.1.1
+---
+* Support configure OpenAPI URI in Spring Cloud Gateway route configs.
+
 1.1.0
 ---
 * Command `az spring create` has new argument "--ingress-read-timeout" to set ingress read timeout when create Azure Spring Apps.